requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-environ==0.4.5
django-js-asset==1.2.2
django-registration-redux==2.9
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Pillow==8.2.0
pip==21.1.2
pkg-resources==0.0.0
python-decouple==3.4
pytz==2021.1
setuptools==57.0.0
sqlparse==0.4.1
